Create abstract class
