+++ src/Generators/AbstractGenerator.php
@@ -1,0 +1,58 @@
+<?php
+
+namespace tiagomichaelsousa\LaravelResources\Generators;
+
+use Illuminate\Support\Facades\File;
+use tiagomichaelsousa\LaravelResources\Exceptions\File as FileException;
+
+abstract class AbstractGenerator implements Generator
+{
+    /**
+     * The model for that will be used in the policy.
+     *
+     * @var string
+     */
+    protected $model;
+
+    /**
+     * Create a new command instance.
+     *
+     * @return void
+     */
+    public function __construct($model)
+    {
+        $this->model = $model;
+    }
+    
+    /**
+     * Verify if the directory and create one if it doesn't.
+     *
+     * @return bool
+     */
+    public function directoryExists($path)
+    {
+        return (bool) File::isDirectory($path) ?: make_directory($path);
+    }
+
+    /**
+     * Verify if the resource already exists.
+     *
+     * @return mixed|\tiagomichaelsousa\LaravelResources\Exceptions\File
+     */
+    public function fileAlreadyExists($path)
+    {
+        if (File::exists($path)) {
+            throw FileException::alreadyExistsInDirectory($path);
+        }
+    }
+    
+    /**
+     * Generate the file name.
+     *
+     * @return string
+     */
+    public function fileName()
+    {
+        return "{$this->className()}.php";
+    }
+}
