+++ src/Generators/MigrationGenerator.php
@@ -3,29 +3,11 @@
 namespace tiagomichaelsousa\LaravelResources\Generators;
 
 use Carbon\Carbon;
-use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
-use tiagomichaelsousa\LaravelResources\Exceptions\File as FileException;
+use Illuminate\Support\Facades\File;
 
-class MigrationGenerator implements Generator
+class MigrationGenerator extends AbstractGenerator
 {
-    /**
-     * The model for that will be used in the policy.
-     *
-     * @var string
-     */
-    private $model;
-
-    /**
-     * Create a new command instance.
-     *
-     * @return void
-     */
-    public function __construct($model)
-    {
-        $this->model = $model;
-    }
-
     /**
      * Get the Stub for the policy.
      *
@@ -47,28 +29,6 @@ class MigrationGenerator implements Generator
             '{{CLASS_NAME}}' => $this->className(),
             '{{TABLE_NAME}}' => Str::lower(Str::snake(Str::plural($this->model))),
         ]);
-    }
-
-    /**
-     * Verify if the resource already exists.
-     *
-     * @return mixed|\tiagomichaelsousa\LaravelResources\Exceptions\File
-     */
-    public function fileAlreadyExists($path)
-    {
-        if (File::exists($path)) {
-            throw FileException::alreadyExistsInDirectory($path);
-        }
-    }
-
-    /**
-     * Verify if the directory and create one if it doesn't.
-     *
-     * @return bool
-     */
-    public function directoryExists($path)
-    {
-        return (bool) File::isDirectory($path) ?: make_directory($path);
     }
 
     /**
