+++ src/Generators/RouteGenerator.php
@@ -2,29 +2,12 @@
 
 namespace tiagomichaelsousa\LaravelResources\Generators;
 
-use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\File;
 use tiagomichaelsousa\LaravelResources\Exceptions\File as FileException;
 
-class RouteGenerator implements Generator
+class RouteGenerator extends AbstractGenerator
 {
-    /**
-     * The model for that will be used in the route.
-     *
-     * @var string
-     */
-    private $model;
-
-    /**
-     * Create a new command instance.
-     *
-     * @return void
-     */
-    public function __construct($model)
-    {
-        $this->model = $model;
-    }
-
     /**
      * Get the Stub for the request.
      *
@@ -57,19 +40,9 @@ class RouteGenerator implements Generator
      */
     public function fileAlreadyExists($path)
     {
-        if (! File::exists($path)) {
+        if (File::missing($path)) {
             throw FileException::doesNotExists($path);
         }
-    }
-
-    /**
-     * Verify if the directory and create one if it doesn't.
-     *
-     * @return bool
-     */
-    public function directoryExists($path)
-    {
-        return (bool) File::isDirectory($path) ?: make_directory($path);
     }
 
     /**
